capture: refuse a Graph sent-folder answer that names no folder
The one comparison the whole T1 gate rests on could read the absence of the
evidence as the evidence. SentFolderID returned ("", nil) whenever Graph
answered 2xx without an id — a stripped $select, a partial-response gateway —
and ListAfter leaves ParentFolderID empty for the same reason, so
`msg.ParentFolderID == sentFolder` compared two empty strings, came back true,
and attested every message on the page whose From matched the owner. The
activity natural key then makes that permanent.

Graph was the only provider that could fail in this direction: Gmail's
hasSentLabel over absent labels yields false, and IMAP's sentAttr yields false
when the server reports no special-use. An id-less response is now a provider
fault, so the page stops and retries from its committed token, and the
comparison refuses an empty id besides.

The rest is the review's craftsmanship pass. The fitness test's compile-time pin
now fixes the signature the walk assumes, not just a name, and no longer claims
to pin something the walk does not use. The direction vocabulary the port now
owns is adopted at its last Go call site — one producer using the constant and
its nearest consumer using a literal is the tell that a second feature does not
read like the first. The port documents the two things unexporting costs: the
attestation reads Direction as it stands at the call, and cannot survive a
serialization boundary. And graph/client.go gives up the transport block, which
belonged to no individual call and had already forced a correct comment to be
shortened to fit the line cap.

Also recorded, per an adversarial review that refuted the other readings: T1
evidence is workspace-scoped by design. Narrowing it to the capturing
connection would suppress a colleague's genuine contact, and any member who can
capture can already create the record directly.
